Split resolver-backed typechecker collection
