+++ docs/PHASE_PLAN.md
@@ -147,6 +147,9 @@ checked-in docs, tests, and commits only.
 - Typechecker resolver behavior-ref metadata collection and restoration now
   lives in a dedicated helper module, keeping resolver-owned impl/requires/
   extends handoff state out of the broad typechecker root.
+- Typechecker resolver-backed declaration collection orchestration now lives in
+  a dedicated helper module, keeping resolver handoff passes out of the broad
+  typechecker root.
 - Resolver validation replay now collects expected declaration symbols,
   expected local symbols, import-validation state, and behavior-association
   replay tasks in one declaration pass before checking resolver-owned extras,
